Move clients to their own folder.
