+++ src/query/useGetSEP10cChallenge.ts
@@ -2,7 +2,7 @@ import { useMutation } from "@tanstack/react-query";
 
 import { SEP10cService } from "@/services/SEP10cService";
 import { GetSEP10cChallengeRequest, GetSEP10cChallengeResponse } from "@/types/types";
-import { SEP10cClientToml } from "@/services/SEP10cClientToml";
+import { SEP10cClientToml } from "@/services/clients/SEP10cClientToml";
 
 export const useGetSEP10cChallenge = () => {
   const mutation = useMutation<GetSEP10cChallengeResponse, Error, GetSEP10cChallengeRequest>({
